# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Water Filter Maintenance Tracker API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface User {
  id: number;
  username: string;
  createdAt: string;
}

export interface AuthInput {
  /** @minLength 3 */
  username: string;
  /** @minLength 6 */
  password: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export type FilterTemplateType = typeof FilterTemplateType[keyof typeof FilterTemplateType];


export const FilterTemplateType = {
  single: 'single',
  three_stage: 'three_stage',
  five_stage: 'five_stage',
  seven_stage: 'seven_stage',
  custom: 'custom',
} as const;

export type FilterStatus = typeof FilterStatus[keyof typeof FilterStatus];


export const FilterStatus = {
  healthy: 'healthy',
  warning: 'warning',
  overdue: 'overdue',
} as const;

export type CartridgeStatus = typeof CartridgeStatus[keyof typeof CartridgeStatus];


export const CartridgeStatus = {
  healthy: 'healthy',
  warning: 'warning',
  overdue: 'overdue',
} as const;

export interface Cartridge {
  id: number;
  filterId: number;
  name: string;
  stageNumber: number;
  /** @nullable */
  lastReplacedDate?: string | null;
  intervalDays: number;
  /** @nullable */
  nextReplacementDate?: string | null;
  status: CartridgeStatus;
  /** @nullable */
  daysRemaining?: number | null;
}

export interface Filter {
  id: number;
  /** @nullable */
  userId?: number | null;
  name: string;
  /** @nullable */
  location?: string | null;
  templateType: FilterTemplateType;
  /** @nullable */
  installationDate?: string | null;
  /** @nullable */
  notes?: string | null;
  status: FilterStatus;
  cartridges?: Cartridge[];
  createdAt: string;
  updatedAt: string;
}

export type FilterInputTemplateType = typeof FilterInputTemplateType[keyof typeof FilterInputTemplateType];


export const FilterInputTemplateType = {
  single: 'single',
  three_stage: 'three_stage',
  five_stage: 'five_stage',
  seven_stage: 'seven_stage',
  custom: 'custom',
} as const;

export interface FilterInput {
  /** @nullable */
  userId?: number | null;
  /** @minLength 1 */
  name: string;
  /** @nullable */
  location?: string | null;
  templateType: FilterInputTemplateType;
  /** @nullable */
  installationDate?: string | null;
  /** @nullable */
  notes?: string | null;
}

export interface FilterUpdate {
  name?: string;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  installationDate?: string | null;
  /** @nullable */
  notes?: string | null;
}

export interface FiltersSummary {
  total: number;
  healthy: number;
  warning: number;
  overdue: number;
  totalCartridges: number;
  overdueCartridges: number;
}

export type UpcomingReplacementStatus = typeof UpcomingReplacementStatus[keyof typeof UpcomingReplacementStatus];


export const UpcomingReplacementStatus = {
  healthy: 'healthy',
  warning: 'warning',
  overdue: 'overdue',
} as const;

export interface UpcomingReplacement {
  filterId: number;
  filterName: string;
  cartridgeId: number;
  cartridgeName: string;
  stageNumber: number;
  nextReplacementDate: string;
  daysRemaining: number;
  status: UpcomingReplacementStatus;
}

export interface CartridgeInput {
  /** @minLength 1 */
  name: string;
  /** @minimum 1 */
  stageNumber: number;
  /** @nullable */
  lastReplacedDate?: string | null;
  /** @minimum 1 */
  intervalDays: number;
}

export interface CartridgeUpdate {
  name?: string;
  stageNumber?: number;
  /** @nullable */
  lastReplacedDate?: string | null;
  intervalDays?: number;
}

export interface ReplaceInput {
  /** @nullable */
  replacedAt?: string | null;
  /** @nullable */
  notes?: string | null;
}

export interface ReplacementRecord {
  id: number;
  filterId: number;
  filterName: string;
  cartridgeId: number;
  cartridgeName: string;
  stageNumber: number;
  replacedAt: string;
  /** @nullable */
  notes?: string | null;
}

export type SettingsLanguage = typeof SettingsLanguage[keyof typeof SettingsLanguage];


export const SettingsLanguage = {
  en: 'en',
  ar: 'ar',
} as const;

export type SettingsTheme = typeof SettingsTheme[keyof typeof SettingsTheme];


export const SettingsTheme = {
  light: 'light',
  dark: 'dark',
  system: 'system',
} as const;

export interface Settings {
  id: number;
  /** @nullable */
  userId?: number | null;
  reminderDays: number[];
  notificationsEnabled: boolean;
  language: SettingsLanguage;
  theme: SettingsTheme;
  themeColor: string;
}

export type SettingsUpdateLanguage = typeof SettingsUpdateLanguage[keyof typeof SettingsUpdateLanguage];


export const SettingsUpdateLanguage = {
  en: 'en',
  ar: 'ar',
} as const;

export type SettingsUpdateTheme = typeof SettingsUpdateTheme[keyof typeof SettingsUpdateTheme];


export const SettingsUpdateTheme = {
  light: 'light',
  dark: 'dark',
  system: 'system',
} as const;

export interface SettingsUpdate {
  reminderDays?: number[];
  notificationsEnabled?: boolean;
  language?: SettingsUpdateLanguage;
  theme?: SettingsUpdateTheme;
  themeColor?: string;
}

export type BannerPosition = typeof BannerPosition[keyof typeof BannerPosition];


export const BannerPosition = {
  home_top: 'home_top',
  home_bottom: 'home_bottom',
  filter_detail: 'filter_detail',
} as const;

export interface Banner {
  id: number;
  position: BannerPosition;
  imageUrl: string;
  /** @nullable */
  clickUrl?: string | null;
  enabled: boolean;
  createdAt: string;
}

export type BannerInputPosition = typeof BannerInputPosition[keyof typeof BannerInputPosition];


export const BannerInputPosition = {
  home_top: 'home_top',
  home_bottom: 'home_bottom',
  filter_detail: 'filter_detail',
} as const;

export interface BannerInput {
  position: BannerInputPosition;
  imageUrl: string;
  /** @nullable */
  clickUrl?: string | null;
  enabled: boolean;
}

export type BannerUpdatePosition = typeof BannerUpdatePosition[keyof typeof BannerUpdatePosition];


export const BannerUpdatePosition = {
  home_top: 'home_top',
  home_bottom: 'home_bottom',
  filter_detail: 'filter_detail',
} as const;

export interface BannerUpdate {
  position?: BannerUpdatePosition;
  imageUrl?: string;
  /** @nullable */
  clickUrl?: string | null;
  enabled?: boolean;
}

export interface AdminPinInput {
  pin: string;
}

export interface AdminTokenResponse {
  token: string;
  valid: boolean;
}

export type ListFiltersParams = {
/**
 * @nullable
 */
userId?: number | null;
};

export type GetUpcomingReplacementsParams = {
withinDays?: number;
};