# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Water Filter Maintenance Tracker API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
...
  /filters/{filterId}/cartridges:
    get:
      operationId: listCartridges
      tags: [cartridges]
      summary: List cartridges for a filter
      parameters:
        - name: filterId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Cartridge"
    post:
      operationId: createCartridge
      tags: [cartridges]
      summary: Add a cartridge to a filter
      parameters:
        - name: filterId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartridgeInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cartridge"
  /filters/{filterId}/cartridges/{cartridgeId}:
    patch:
      operationId: updateCartridge
      tags: [cartridges]
      summary: Update a cartridge
      parameters:
        - name: filterId
          in: path
          required: true
          schema:
            type: integer
        - name: cartridgeId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartridgeUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cartridge"
    delete:
      operationId: deleteCartridge
      tags: [cartridges]
      summary: Delete a cartridge
      parameters:
        - name: filterId
          in: path
          required: true
          schema:
            type: integer
        - name: cartridgeId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: No content
  /filters/{filterId}/cartridges/{cartridgeId}/replace:
    post:
      operationId: replaceCartridge
      tags: [cartridges]
      summary: Mark a cartridge as replaced
      parameters:
        - name: filterId
          in: path
          required: true
          schema:
            type: integer
        - name: cartridgeId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReplaceInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cartridge"
...
  /admin/verify-pin:
    post:
      operationId: verifyAdminPin
      tags: [banners]
      summary: Verify admin PIN
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminPinInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminTokenResponse"
        "401":
          description: Unauthorized
  /admin/banners:
    get:
      operationId: listAdminBanners
      tags: [banners]
      summary: List all banners (admin)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Banner"
    post:
      operationId: createBanner
      tags: [banners]
      summary: Create a banner
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BannerInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Banner"
  /admin/banners/{bannerId}:
    patch:
      operationId: updateBanner
      tags: [banners]
      summary: Update a banner
      parameters:
        - name: bannerId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BannerUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Banner"
    delete:
      operationId: deleteBanner
      tags: [banners]
      summary: Delete a banner
      parameters:
        - name: bannerId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: No content
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        createdAt:
          type: string
      required: [id, username, createdAt]
    AuthInput:
      type: object
      properties:
        username:
          type: string
          minLength: 3
        password:
          type: string
          minLength: 6
      required: [username, password]
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required: [user, token]
...
    Cartridge:
      type: object
      properties:
        id:
          type: integer
        filterId:
          type: integer
        name:
          type: string
        stageNumber:
          type: integer
        lastReplacedDate:
          type: ["string", "null"]
        intervalDays:
          type: integer
        nextReplacementDate:
          type: ["string", "null"]
        status:
          type: string
          enum: [healthy, warning, overdue]
        daysRemaining:
          type: ["integer", "null"]
      required: [id, filterId, name, stageNumber, intervalDays, status]
    CartridgeInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        stageNumber:
          type: integer
          minimum: 1
        lastReplacedDate:
          type: ["string", "null"]
        intervalDays:
          type: integer
          minimum: 1
      required: [name, stageNumber, intervalDays]
    CartridgeUpdate:
      type: object
      properties:
        name:
          type: string
        stageNumber:
          type: integer
        lastReplacedDate:
          type: ["string", "null"]
        intervalDays:
          type: integer
    ReplaceInput:
      type: object
      properties:
        replacedAt:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
...
    Banner:
      type: object
      properties:
        id:
          type: integer
        position:
          type: string
          enum: [home_top, home_bottom, filter_detail]
        imageUrl:
          type: string
        clickUrl:
          type: ["string", "null"]
        enabled:
          type: boolean
        createdAt:
          type: string
      required: [id, position, imageUrl, enabled, createdAt]
    BannerInput:
      type: object
      properties:
        position:
          type: string
          enum: [home_top, home_bottom, filter_detail]
        imageUrl:
          type: string
        clickUrl:
          type: ["string", "null"]
        enabled:
          type: boolean
      required: [position, imageUrl, enabled]
    BannerUpdate:
      type: object
      properties:
        position:
          type: string
          enum: [home_top, home_bottom, filter_detail]
        imageUrl:
          type: string
        clickUrl:
          type: ["string", "null"]
        enabled:
          type: boolean
    AdminPinInput:
      type: object
      properties:
        pin:
          type: string
      required: [pin]
    AdminTokenResponse:
      type: object
      properties:
        token:
          type: string
        valid:
          type: boolean
      required: [token, valid]