# lib/api-zod/src/generated/types/banner.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Water Filter Maintenance Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { BannerPosition } from './bannerPosition';

export interface Banner {
  id: number;
  position: BannerPosition;
  imageUrl: string;
  /** @nullable */
  clickUrl?: string | null;
  enabled: boolean;
  createdAt: string;
}