# lib/api-zod/src/generated/types/cartridge.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Water Filter Maintenance Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { CartridgeStatus } from './cartridgeStatus';

export interface Cartridge {
  id: number;
  filterId: number;
  name: string;
  stageNumber: number;
  /** @nullable */
  lastReplacedDate?: string | null;
  intervalDays: number;
  /** @nullable */
  nextReplacementDate?: string | null;
  status: CartridgeStatus;
  /** @nullable */
  daysRemaining?: number | null;
}