# lib/api-zod/src/generated/types/cartridgeUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Water Filter Maintenance Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface CartridgeUpdate {
  name?: string;
  stageNumber?: number;
  /** @nullable */
  lastReplacedDate?: string | null;
  intervalDays?: number;
}